requirements.txt
Appium-Python-Client==1.1.0
uiautomator2==2.15.0
weditor==0.6.3



